Reorganise algorithm folder
