+++ src/main/scala/com/raphtory/algorithms/generic/centrality/WeightedDegree.scala
@@ -1,4 +1,4 @@
-package com.raphtory.algorithms
+package com.raphtory.algorithms.generic.centrality
 
 import com.raphtory.core.model.algorithm.{GraphAlgorithm, GraphPerspective, Row, Table}
